+++ src/command_changeset_filter.cpp
@@ -159,7 +159,7 @@ bool CommandChangesetFilter::run() {
     header.set("generator", m_generator);
 
     m_vout << "Opening output file...\n";
-    osmium::io::Writer writer(m_output_file, header, m_output_overwrite);
+    osmium::io::Writer writer(m_output_file, header, m_output_overwrite, m_fsync);
     auto out = osmium::io::make_output_iterator(writer);
 
     m_vout << "Filtering data...\n";