+++ src/command_merge_changes.cpp
@@ -104,7 +104,7 @@ bool CommandMergeChanges::run() {
     header.set("generator", m_generator);
 
     m_vout << "Opening output file...\n";
-    osmium::io::Writer writer(m_output_file, header, m_output_overwrite);
+    osmium::io::Writer writer(m_output_file, header, m_output_overwrite, m_fsync);
     auto out = osmium::io::make_output_iterator(writer);
 
     // Now we sort all objects and write them in order into the